+++ test-website/src/Pages/Test/Components/Components.tsx
@@ -62,7 +62,9 @@ function MyActionDialog(props: MyActionDialogProps) {
             title="Action Dialog"
             actionButtonText="OK"
             actionInProgress={actionInProgress}
-            action={onClose}
+            action={() => {
+                closeRef.current()
+            }}
             showFooter={showFooter}
             onClose={onClose}
             onCancel={
@@ -390,6 +392,7 @@ export default class Page extends React.Component<PageProps, PageState> {
                             >
                                 Standalone confirm dialog
                             </button>
+
                             <button
                                 className="btn btn-danger"
                                 onClick={() =>
@@ -397,14 +400,12 @@ export default class Page extends React.Component<PageProps, PageState> {
                                         errorDialogArgs: {},
                                     })
                                 }
+                                data-tooltip="This is to test that the modal gets fully removed if it
+                                hits an error in componentDidMount(). Make sure you can
+                                scroll."
                             >
                                 Dialog error test
                             </button>
-                            <span>
-                                This is to test that the modal gets fully removed if it
-                                hits an error in componentDidMount(). Make sure you can
-                                scroll.
-                            </span>
                         </div>
                     </div>
                 </div>